Move show_content_type_selection_widget to feincms_admin_tags, that it where it belongs
