+++ feincms/templates/admin/feincms/content_editor.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load feincms_tags %}
+{% load feincms_admin_tags %}
 <div id="main_wrapper">
     {% for region in original.template.regions %}
         <div class="navi_tab" id="{{ region.key }}_tab">{{ region.title }}</div>
